lavc: change type of AVFrame.channels field from int64_t to int
"channels" is an int in all the other places in the libraries, and the
av_frame_*_channels() accessors return and set an int, so this should not
implicate ABI breaks.

+++ libavcodec/avcodec.h
@@ -1501,7 +1501,7 @@ typedef struct AVFrame {
      * - encoding: unused
      * - decoding: Read by user.
      */
-    int64_t channels;
+    int channels;
 
     /**
      * size of the corresponding packet containing the compressed
